eslint完善

+++ .eslintrc.js
@@ -12,7 +12,7 @@ module.exports = {
     ],
     "rules": {
         "indent": [2, 4],
-        "quotes": [2, "double"],
+        "quotes": [2, "single"],
         "semi": [2, "always"]
     }
 };
